Added Edit profile pic feature with Data module.
