+++ index.html
@@ -169,7 +169,7 @@
                 <!-- <iframe class="widget-iframe" src="https://www.youtube.com/embed/D_uLM5i0Z4c" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe> -->
             </aside>
 
-            <script src="../script/homepage.js"></script>
+            <script type="module" src="../script/homepage.js"></script>
         </section>
     </body>
 </html>
